+++ src/main/java/dev/qwe664/bbc/placeholder/BBCExpansion.java
@@ -58,7 +58,7 @@ public class BBCExpansion extends PlaceholderExpansion {
 
     @Override
     public String getVersion() {
-        return plugin.getDescription().getVersion();
+        return plugin.getPluginMeta().getVersion();
     }
 
     /**